+++ CHANGELOG.md
@@ -21,3 +21,7 @@
 # Version 1.0.0
 - Initial release of the modular plugin.
     - See [CHANGELOG](https://github.com/dpa99c/cordova-plugin-firebasex/blob/master/CHANGELOG.md) for the full changelog of the main plugin which includes this module as a dependency.
+# Tipout changes
+
+- (android) add an explicit app notification-settings redirect with an
+  application-details fallback on Android 7.x.